pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "memory*", "audio*", "training*", "api*", "demo*"]

[project]
name = "brandmind"
version = "1.0.0"
description = "Autonomous Brand Voice Agent — AI Agent Economy Hackathon 2026"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.75.0",
    "mem0ai==2.0.1",
    "fish-audio-sdk==1.3.0",
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.0",
    "httpx==0.28.1",
    "beautifulsoup4==4.13.4",
    "requests==2.32.3",
    "python-dotenv==1.1.0",
    "rich==13.9.4",
    "typer==0.15.2",
    "pydantic==2.11.3",
    "x402[httpx,evm]==2.8.0",
]
